feat: implement LocaleSelector component for language selection in Navbar
